feat: lod 和 间接光镜面反射

+++ Assets/BXRenderPipline/Shaders/Libiary/UnityInput.hlsl
@@ -60,7 +60,7 @@ CBUFFER_START(UnityPerDraw)
 
     half4 unity_ProbesOcclusion;
 
-    float4 unity_SpecCube0_HDR;
+    half4 unity_SpecCube0_HDR;
 
     float4 unity_LightmapST;
     float4 unity_DynamicLightmapST;
